split handlers & controllers. Move application logic to controllers and make them swappable
